Added new files (   src/Placeholders.php,   src/Route.php,   src/RouteBuilder.php,   src/RouteCollection.php,   src/Router.php )
Class `Placeholders` contains Routing built-in params regex patterns
to extract parameters to be passed for the controller action (method)

Class `Route` will be the linker between request URL and Controller
actions with additional info, like parameters to pass when invoking
Controller action.

Class `RouteBuilder` will be building and create `Route `this is the
entry point of using `Zelasli/Routing` module, also will handle
grouping related `Route`

Class `RouteCollection` is a repository where all created `Route` will
be stored in.

Class `Router` this is the core of this module that will do it's job
to find `Route`.